fix issue with source discretization in backend by adding an optional step in discretize_inds which slightly enlarges the size of boxes
